+++ fiber/chain_interactions/models.py
@@ -1,4 +1,5 @@
-from typing import TypedDict
+from enum import Enum
+from typing import TypedDict, TypeAlias
 
 from cryptography.fernet import Fernet
 from pydantic import BaseModel
@@ -27,3 +28,26 @@ class Node(BaseModel):
 class ParamWithTypes(TypedDict):
     name: str
     type: str
+
+
+class CommitmentDataFieldType(Enum):
+    RAW = "Raw"
+    BLAKE_TWO_256 = "BlakeTwo256"
+    SHA_256 = "Sha256"
+    KECCAK_256 = "Keccak256"
+    SHA_THREE_256 = "ShaThree256"
+
+
+CommitmentDataField: TypeAlias = tuple[CommitmentDataFieldType, bytes] | None
+
+
+class CommitmentQuery(BaseModel):
+    fields: list[CommitmentDataField]
+    block: int
+    deposit: int
+
+
+class RawCommitmentQuery(BaseModel):
+    data: bytes
+    block: int
+    deposit: int
