+++ fiber/constants.py
@@ -9,6 +9,8 @@ FINNEY_TEST_NETWORK = "test"
 FINNEY_SUBTENSOR_ADDRESS = "wss://entrypoint-finney.opentensor.ai:443"
 FINNEY_TEST_SUBTENSOR_ADDRESS = "wss://test.finney.opentensor.ai:443/"
 
+EMPTY_COMMITMENT_FIELD_TYPE = "None"
+
 SUBTENSOR_NETWORK_TO_SUBTENSOR_ADDRESS = {
     FINNEY_NETWORK: FINNEY_SUBTENSOR_ADDRESS,
     FINNEY_TEST_NETWORK: FINNEY_TEST_SUBTENSOR_ADDRESS,
